added tft support, fixed a nasty bug in snake library related to the data size of the buffer queue, moved from byte to double byte

+++ lib/Snake/Snake.cpp
@@ -1,6 +1,6 @@
 /*
 Snake.cpp Library implementation
-Created by Gabriele Tazzari 
+Created by Gabriele Tazzari
 */
 
 #include "Arduino.h"
@@ -20,7 +20,7 @@ Snake::Snake (uint16_t w, uint16_t h)
   todely=0;
   _direction='r'; //initial direction
   uint16_t bytes = w*h;
-  if((buffer = (uint8_t *)malloc(bytes))) {
+  if((buffer = (uint16_t *)malloc(bytes))) {
       memset(buffer, 0, bytes);
   }
   buffer[0]=posx+posy*_width; //set starting position
@@ -70,11 +70,12 @@ void Snake::nextmove()
     }
   }
   todelx=buffer[length-1] % _width;
-  todely=buffer[length-1] /  _width;
+  todely=buffer[length-1] / _width;
   //delete last pixel at the end of cycle
   buffer[length-1]=0;
   //remove last pixel from array
   buffer[0]=currentpos; //base point set
+
   //insert new position into array head
 }
 Snake::~Snake(void) {
